# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the BrightBund social platform with economy, maps, chat, and gamification",
        "title": "BrightBund API",
        "termsOfService": "https://brightbund.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@brightbund.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://brightbund.com/license"
        },
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
...
        "/profiles/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve the authenticated user's profile. Auto-creates if not exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get my profile",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete the authenticated user's profile. Profile can be auto-recreated on next access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete my profile",
                "responses": {
                    "204": {
                        "description": "Profile deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Profile deletion failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the authenticated user's profile information. Supports partial updates (PATCH semantics) - only send fields you want to update. Auto-populates location if IP provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Update my profile",
                "parameters": [
                    {
                        "description": "Profile update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profiles.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated profile",
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Profile update failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/profiles/me/allies": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get list of users following the authenticated user.",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get my allies (my subscribers)",
                "responses": {
                    "200": {
                        "description": "List of allies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profiles.AllyProfile"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        },
        "/profiles/me/avatar": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload an avatar image for the authenticated user. Max 5MB, supports jpg/png/webp.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Upload avatar image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated profile with new avatar URL",
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    },
                    "400": {
                        "description": "File required, too large, or invalid type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Upload failed or storage not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/profiles/ranks": {
            "get": {
                "description": "Retrieve a list of all available ranks with their properties, levels, and seal requirements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get all ranks",
                "responses": {
                    "200": {
                        "description": "List of all ranks",
                        "schema": {
                            "$ref": "#/definitions/ranks.RankListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "profiles.AllyProfile": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "rank_tier": {
                    "description": "Computed",
                    "type": "string"
                },
                "reputation_score": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "ranks.RankListResponse": {
            "type": "object",
            "properties": {
                "ranks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ranks.RankWithLevels"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and your JWT Access Token (not UUID). Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}